Created BiomeMap and implemented biome blending (with lots of artifacts)

+++ header/Mesh.h
@@ -11,7 +11,7 @@ struct Vertex
 	glm::vec3 position;
 	glm::vec3 normal{ 0.0f, 1.0f, 0.0f };
 	glm::vec3 lodColor{ 0.0f };
-	int biome{ 0 };
+	float biome{ 0 };
 };
 
 class Mesh
